+++ test/common/multi_polygon_tests.rb
@@ -57,18 +57,12 @@ module RGeo
         end
 
         def test_creation_overlapping
-          assert_raises(RGeo::Error::InvalidGeometry) do
-            @factory.multi_polygon([@poly1, @poly1])
-          end
-          geom = @lenient_factory.multi_polygon([@poly1, @poly1])
+          geom = @factory.multi_polygon([@poly1, @poly1])
           assert_equal RGeo::Feature::MultiPolygon, geom.geometry_type
         end
 
         def test_creation_connected
-          assert_raises(RGeo::Error::InvalidGeometry) do
-            @factory.multi_polygon([@poly3, @poly4])
-          end
-          geom = @lenient_factory.multi_polygon([@poly3, @poly4])
+          geom = @factory.multi_polygon([@poly3, @poly4])
           assert_equal RGeo::Feature::MultiPolygon, geom.geometry_type
         end