+++ test/geos_capi/validity_test.rb
@@ -12,6 +12,6 @@ class GeosValidityTest < Minitest::Test # :nodoc:
   include RGeo::Tests::Common::ValidityTests
 
   def setup
-    @factory = RGeo::Geos.factory(uses_lenient_multi_polygon_assertions: true)
+    @factory = RGeo::Geos.factory
   end
 end if RGeo::Geos.capi_supported?